refactor: 마우스오버핸들러 함수를 검색결과 컴포넌트로 이동
